Separate history filtering from results
