+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "zadiag-pwa",
   "private": true,
-  "version": "0.11.29",
+  "version": "0.11.30",
   "type": "module",
   "packageManager": "pnpm@11.7.0",
   "engines": {
