+++ src/components/RoutineHistoryPanel.tsx
@@ -1,13 +1,12 @@
-import { useEffect, useMemo, useState, type CSSProperties } from 'react';
+import { useMemo, useState, type CSSProperties } from 'react';
 import type { Locale, RoutineAssignment, VerificationEvent, VerificationStatus } from '../domain/models';
 import type { MessageKey } from '../services/i18n';
 import { presentRoutine } from '../domain/routinePresentation';
 import { AppIcon, routineIconName } from './Icon';
-import { StatusPill, statusMessageKey } from './StatusPill';
-import { canRetakeCapture, isSuccessfulVerification, stalePendingCheckReason, withResolvedEventStatuses } from '../domain/adherence';
+import { StatusPill } from './StatusPill';
+import { canRetakeCapture, stalePendingCheckReason, withResolvedEventStatuses } from '../domain/adherence';
 import { coalesceActivePendingEventsByRoutine } from '../domain/dashboardChecks';
 import { EmptyState, ListRow } from './ui';
-import { readUiStorageJson, writeUiStorageString } from '../services/uiStorage';
 import { languageTag } from '../services/locale';
 
 const eventTimestamp = (event: VerificationEvent) =>
@@ -33,104 +32,11 @@ const hiddenReasonCodes = new Set(['analysis_unavailable', 'self_validated']);
 const displayReason = (reason?: string) =>
   reason && !hiddenReasonCodes.has(reason) ? reason : undefined;
 
-const historyFilterStorageKey = (titleId: string) => `zadiag.historyFilters.${titleId}`;
-
-const readStoredFilters = (titleId: string) => {
-  const empty = { statuses: [] as VerificationStatus[], routineIds: [] as string[] };
-  return readUiStorageJson(historyFilterStorageKey(titleId), empty, (value) => {
-    const parsed = value as Partial<{ statuses: VerificationStatus[]; routineIds: string[] }>;
-    return {
-      statuses: Array.isArray(parsed.statuses) ? parsed.statuses : [],
-      routineIds: Array.isArray(parsed.routineIds) ? parsed.routineIds : [],
-    };
-  });
-};
-
 const analysisTag = (event: VerificationEvent, locale: Locale) => {
   if (event.analysisSource === 'ai') return locale === 'fr' ? 'IA' : 'AI';
   if (event.analysisSource === 'self' || event.reason === 'self_validated') return 'Auto';
   return undefined;
 };
-
-export const groupedVerificationStatuses = (statuses: VerificationStatus[]) => {
-  const groups = new Map<VerificationStatus, VerificationStatus[]>();
-  statuses.forEach((eventStatus) => {
-    const status = isSuccessfulVerification({ status: eventStatus }) ? 'detected' : eventStatus;
-    groups.set(status, [...(groups.get(status) ?? []), eventStatus]);
-  });
-  return Array.from(groups, ([status, eventStatuses]) => ({ status, eventStatuses }));
-};
-
-export function useHistoryFilters(titleId: string) {
-  const [excludedStatuses, setExcludedStatuses] = useState<VerificationStatus[]>(() => readStoredFilters(titleId).statuses);
-  const [excludedRoutineIds, setExcludedRoutineIds] = useState<string[]>(() => readStoredFilters(titleId).routineIds);
-  useEffect(() => {
-    writeUiStorageString(historyFilterStorageKey(titleId), JSON.stringify({
-      statuses: excludedStatuses,
-      routineIds: excludedRoutineIds,
-    }));
-  }, [excludedRoutineIds, excludedStatuses, titleId]);
-  return {
-    excludedStatuses,
-    excludedRoutineIds,
-    toggleRoutine: (routineId: string) => setExcludedRoutineIds((current) =>
-      current.includes(routineId) ? current.filter((item) => item !== routineId) : [...current, routineId]),
-    toggleStatuses: (statuses: VerificationStatus[]) => setExcludedStatuses((current) => {
-      const allActive = statuses.every((status) => !current.includes(status));
-      return allActive
-        ? Array.from(new Set([...current, ...statuses]))
-        : current.filter((status) => !statuses.includes(status));
-    }),
-  };
-}
-
-export function HistoryFilterControls({
-  assignments,
-  events,
-  locale,
-  excludedRoutineIds,
-  excludedStatuses,
-  onToggleRoutine,
-  onToggleStatuses,
-  t,
-}: {
-  assignments: RoutineAssignment[];
-  events: VerificationEvent[];
-  locale: Locale;
-  excludedRoutineIds: string[];
-  excludedStatuses: VerificationStatus[];
-  onToggleRoutine: (routineId: string) => void;
-  onToggleStatuses: (statuses: VerificationStatus[]) => void;
-  t: (key: MessageKey) => string;
-}) {
-  const statuses = groupedVerificationStatuses(Array.from(new Set(
-    withResolvedEventStatuses(coalesceActivePendingEventsByRoutine(events, Date.now()), Date.now())
-      .map((event) => event.status),
-  )));
-  return (
-    <div className="history-filter-controls">
-      <div className="filter-group">
-        <span>{t('filterByRoutine')}</span>
-        <div className="filter-chips">
-          {assignments.map((assignment) => {
-            const visual = presentRoutine(assignment.routine, locale);
-            const active = !excludedRoutineIds.includes(assignment.routineId);
-            return <button type="button" key={assignment.id} aria-pressed={active} className={active ? 'active' : ''} onClick={() => onToggleRoutine(assignment.routineId)}>{visual.name}</button>;
-          })}
-        </div>
-      </div>
-      <div className="filter-group">
-        <span>{t('filterByStatus')}</span>
-        <div className="filter-chips">
-          {statuses.map(({ status, eventStatuses }) => {
-            const active = eventStatuses.every((eventStatus) => !excludedStatuses.includes(eventStatus));
-            return <button type="button" key={status} aria-pressed={active} className={`filter-status-${status} ${active ? 'active' : ''}`} onClick={() => onToggleStatuses(eventStatuses)}>{t(statusMessageKey(status))}</button>;
-          })}
-        </div>
-      </div>
-    </div>
-  );
-}
 
 export function RoutineHistoryPanel({
   assignments,
@@ -142,7 +48,6 @@ export function RoutineHistoryPanel({
   onOpenEvent,
   onRequestCheck,
   participants,
-  participantForEvent,
   colorForEvent,
   excludedParticipantIds = [],
   excludedRoutineIds,
@@ -159,7 +64,6 @@ export function RoutineHistoryPanel({
   onOpenEvent?: (event: VerificationEvent) => void;
   onRequestCheck?: (routineId: string) => Promise<void>;
   participants?: Array<{ id: string; displayName: string; profileColor: string }>;
-  participantForEvent?: (event: VerificationEvent) => { id: string; displayName: string; profileColor: string } | undefined;
   colorForEvent?: (event: VerificationEvent) => string | undefined;
   excludedParticipantIds?: string[];
   excludedRoutineIds: string[];
